Changes to support separate controllers for each tab
Added newcontroller for the table tab, added new fxml file for the table, edited main and map fxml files to include the extra fxml files and link to controllers, and moved table methods and fxml declarations from maincontroller to the table controller. Slight edits of remaining code in main controller to make it tidier.