Added Graph tab in main gui
moved some of the graphing functionality to the backend, created a graphcontroller, added a checkcombobox to select wards/crime types etc. Having trouble giving it a max number to select but working on a solution to this. Graphs should be shown in the graph tab when create graph button is clicked but seems to be a null pointer exception either for graphcontroller or graphborderpane depending on whether or not graphcontroller is created explicitly in maincontroller